requirements.txt
# Python dependencies for GitHub Code Review MCP Server

# MCP Framework
mcp>=1.0.0

# GitHub API Integration
PyGithub>=2.1.1
github3.py>=4.0.1

# Code Analysis Tools
ast-tools>=0.1.8
tree-sitter>=0.20.0

# Security Analysis
bandit>=1.7.5
safety>=2.3.5
semgrep>=1.45.0

# Code Quality Analysis
pylint>=3.0.3
flake8>=6.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1

# HTTP and API
requests>=2.31.0
aiohttp>=3.9.1
httpx>=0.25.2

# Data Processing
pydantic>=2.5.0
pyyaml>=6.0.1
toml>=0.10.2

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
pathlib-mate>=1.3.2

# Testing (development)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Optional: Machine Learning for Advanced Analysis
scikit-learn>=1.3.2
numpy>=1.24.4
pandas>=2.1.4

# Optional: Natural Language Processing
transformers>=4.36.0
torch>=2.1.1
